requirements.txt
paho-mqtt
pymavlink
customtkinter
tkintermapview
websockets
opencv-python